Extract rollback guard module
