+++ src/cstr.rs
@@ -67,7 +67,8 @@ use std::collections::{
 };
 use std::ffi::CStr;
 
-use crate::internal::{Interned, VecEntryRollbackGuard};
+use crate::internal::Interned;
+use crate::rollback::VecEntryRollbackGuard;
 use crate::{DEFAULT_SYMBOL_TABLE_CAPACITY, Symbol, SymbolOverflowError};
 
 /// An iterator over all [`Symbol`]s in a [`SymbolTable`].
@@ -776,7 +777,7 @@ where
             let slice = unsafe { name.last().as_static_slice() };
 
             self.map.insert(slice, id);
-            name.disarm();
+            name.defuse();
             slice
         };
