+++ src/path.rs
@@ -68,7 +68,8 @@ use std::collections::{
 };
 use std::path::Path;
 
-use crate::internal::{Interned, VecEntryRollbackGuard};
+use crate::internal::Interned;
+use crate::rollback::VecEntryRollbackGuard;
 use crate::{DEFAULT_SYMBOL_TABLE_CAPACITY, Symbol, SymbolOverflowError};
 
 /// An iterator over all [`Symbol`]s in a [`SymbolTable`].
@@ -779,7 +780,7 @@ where
             let slice = unsafe { name.last().as_static_slice() };
 
             self.map.insert(slice, id);
-            name.disarm();
+            name.defuse();
             slice
         };
